# docs/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /features/{id}:
    get:
      operationId: findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureResponseDto"
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
      security: *a3
      summary: Get a Feature by id
      tags: *a4
...
  /features/{featureId}/release-log:
    get:
      operationId: get
      parameters:
        - name: featureId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureReleaseLogResponseDto"
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
      security: &a17
        - access-token: []
      summary: Get release log for a Feature
      tags: &a18
        - release-log
    put:
      operationId: put
      parameters:
        - name: featureId
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertFeatureReleaseLogDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureReleaseLogResponseDto"
        "400":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
      security: *a17
      summary: Create or update release log
      tags: *a18
  /features/{featureId}/bugs:
    get:
      operationId: list
      parameters:
        - name: featureId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureBugTriageResponseDto"
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
      security: &a19
        - access-token: []
      summary: List bug triage rows for a Feature
      tags: &a20
        - bugs
    post:
      operationId: create
      parameters:
        - name: featureId
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeatureBugTriageDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureBugTriageResponseDto"
        "400":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
      security: *a19
      summary: Add a bug triage row
      tags: *a20
  /features/{featureId}/bugs/{bugId}:
    patch:
      operationId: update
      parameters:
        - name: featureId
          required: true
          in: path
          schema:
            type: number
        - name: bugId
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeatureBugTriageDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureBugTriageResponseDto"
        "400":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFailureEnvelopeDto"
      security: *a19
      summary: Update a bug triage row
      tags: *a20
info:
  title: Scoutbook API
  description: Scoutbook backend — structured decision records for PO/PM/Developer/QA teams. Auth uses
    JWT bearer tokens (1h). See docs/rfcs/0001-authentication.md.
  version: 0.1.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Paste the accessToken from POST /auth/login
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        error:
          type: string
          example: Bad Request
        code:
          type: string
          example: VALIDATION_ERROR
          enum:
            - BAD_REQUEST
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
            - HTTP_ERROR
        message:
          type: string
          example: Validation failed
          description: Single human-readable message for UI display
        details:
          example:
            - title must be longer than or equal to 3 characters
          type: array
          items:
            type: string
        path:
          type: string
          example: /features
        timestamp:
          type: string
          example: 2026-09-15T07:00:00.000Z
        requestId:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      required:
        - statusCode
        - error
        - code
        - message
        - path
        - timestamp
    ApiFailureEnvelopeDto:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: "#/components/schemas/ErrorResponseDto"
      required:
        - success
        - error
...
    FeatureResponseDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        title:
          type: string
          example: GitHub Docs Sync
        problem:
          type: string
          example: Decisions never land in the target repo docs/ folder.
        riskTier:
          type: string
          enum:
            - P1
            - P2
            - P3
          example: P2
        currentStage:
          type: string
          enum:
            - IDEA
            - SCOUTING
            - RFC
            - RACI
            - IMPLEMENTATION
            - TESTING
            - REVIEW
            - RELEASE
          example: IDEA
        createdByUserId:
          type: number
          example: 1
        createdAt:
          type: string
          example: 2026-09-15T06:00:00.000Z
        updatedAt:
          type: string
          example: 2026-09-15T06:00:00.000Z
      required:
        - id
        - title
        - problem
        - riskTier
        - currentStage
        - createdByUserId
        - createdAt
        - updatedAt
...
    FeatureReleaseLogResponseDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        featureId:
          type: number
          example: 1
        status:
          type: string
          enum: &a27
            - NOT_STARTED
            - SHIPPED
            - OBSERVING
            - STABLE
          example: SHIPPED
        summary:
          type: string
          example: Shipped to staging; watching error rates
        watchStarted:
          type: boolean
          example: true
        notes:
          type: string
          example: ""
        updatedByUserId:
          type: object
          example: 1
          nullable: true
        createdAt:
          type: string
          example: 2026-09-30T10:00:00.000Z
        updatedAt:
          type: string
          example: 2026-09-30T10:00:00.000Z
      required:
        - id
        - featureId
        - status
        - summary
        - watchStarted
        - notes
        - updatedByUserId
        - createdAt
        - updatedAt
    UpsertFeatureReleaseLogDto:
      type: object
      properties:
        status:
          type: string
          enum: *a27
          example: SHIPPED
        summary:
          type: string
          example: Shipped to staging; watching error rates
          maxLength: 500
        watchStarted:
          type: boolean
          example: true
        notes:
          type: string
          example: Rollback plan reviewed
          maxLength: 500
      required:
        - status
        - watchStarted
    FeatureBugTriageResponseDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        featureId:
          type: number
          example: 1
        whatHappened:
          type: string
          example: Assign button does nothing on mobile
        expected:
          type: string
          example: Assign opens the user picker
        reproduce:
          type: string
          example: 1. Open on phone 2. Tap Assign
        bugType:
          type: string
          enum: &a28
            - REGRESSION
            - SCOUTING_GAP
            - UNSURE
          example: REGRESSION
        riskTier:
          type: string
          enum: &a29
            - P1
            - P2
            - P3
            - UNKNOWN
          example: P2
        status:
          type: string
          enum: &a30
            - OPEN
            - RESOLVED
            - ESCALATED_TO_PO
          example: OPEN
        createdByUserId:
          type: object
          example: 1
          nullable: true
        createdAt:
          type: string
          example: 2026-09-30T10:00:00.000Z
        updatedAt:
          type: string
          example: 2026-09-30T10:00:00.000Z
      required:
        - id
        - featureId
        - whatHappened
        - expected
        - reproduce
        - bugType
        - riskTier
        - status
        - createdByUserId
        - createdAt
        - updatedAt
    CreateFeatureBugTriageDto:
      type: object
      properties:
        whatHappened:
          type: string
          example: Assign button does nothing on mobile
        expected:
          type: string
          example: Assign opens the user picker
        reproduce:
          type: string
          example: |-
            1. Open feature on phone
            2. Tap Assign
          maxLength: 500
        bugType:
          type: string
          enum: *a28
          example: REGRESSION
        riskTier:
          type: string
          enum: *a29
          example: P2
        status:
          type: string
          enum: *a30
          example: OPEN
      required:
        - whatHappened
        - expected
        - bugType
    UpdateFeatureBugTriageDto:
      type: object
      properties:
        whatHappened:
          type: string
          example: Assign button does nothing on mobile
        expected:
          type: string
          example: Assign opens the user picker
        reproduce:
          type: string
          maxLength: 500
        bugType:
          type: string
          enum: *a28
        riskTier:
          type: string
          enum: *a29
        status:
          type: string
          enum: *a30